requirements.txt
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
install==1.3.5
pycodestyle==2.10.0
pyfiglet==0.8.post1
python-dotenv==0.21.0
requests==2.28.1
termcolor==2.1.1
tomli==2.0.1
urllib3==1.26.13
